Implemented saving the encrypted/decrypted text.
Current bugs: ctrl+d kills the program and stops the function from printing the input text.

The input args does not currently check the type of the input parameters such as the key value. This is probably necessary to implement but it is not clear how to do this at this moment.